Moved tests to new class; partial fixes
